1.6 Книги. Синхронизация и работа с ними в целом

+++ Курсач.Core/DB/DatabaseManager.cs
@@ -25,6 +25,7 @@ namespace Курсач.Core.DB
             var dbPath = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), "DB.sqlite");
             Database = new SQLiteAsyncConnection(dbPath);
             Database.CreateTableAsync<User>().Wait();
+            Database.CreateTableAsync<Book>().Wait();
         }
         public async Task<User> GetUserAsync()
         {
@@ -33,13 +34,50 @@ namespace Курсач.Core.DB
 
         public async Task AddUserAsync(User user)
         {
-            await DeleteUserAsync();
-            await Database.InsertAsync(user);
+            if (Database.Table<User>() == null || Database.Table<User>().FirstOrDefaultAsync(b => b.Id == user.Id) == null) // Если в базе нет пользователя или нет именно этого пользователя
+            {
+                await DeleteUserAsync();
+                await DeleteAllBooksAsync();
+                await Database.InsertAsync(user);
+            }
         }
 
         public async Task DeleteUserAsync()
         {
             await Database.DeleteAllAsync<User>();
+        }
+
+        public async Task<Book> GetBookAsync(int id)
+        {
+            return await Database.Table<Book>().FirstOrDefaultAsync(b => b.Id == id);
+        }
+
+        public async Task AddBookAsync(Book book)
+        {
+            await Database.InsertAsync(book);
+        }
+        public async Task UpdateBookAsync(Book book)
+        {
+            await Database.UpdateAsync(book);
+        }
+        public async Task DeleteBookAsync(int id)
+        {
+            var bookToDelete = await Database.Table<Book>().FirstOrDefaultAsync(b => b.Id == id);
+
+            if (bookToDelete != null)
+            {
+                await Database.DeleteAsync(bookToDelete);
+            }
+        }
+
+        public async Task DeleteAllBooksAsync()
+        {
+            await Database.DeleteAllAsync<Book>();
+        }
+
+        public async Task<List<Book>> GetAllBooksAsync()
+        {
+            return await Database.Table<Book>().ToListAsync();
         }
     }
